replace modal with dialog
